Remove the protocol contributor changes
