+++ radix-engine-interface/src/blueprints/pool/multi_resource_pool/invocations.rs
@@ -28,26 +28,6 @@ define_invocation! {
 
 define_invocation! {
     blueprint_name: MultiResourcePool,
-    function_name: instantiate_with_contributor_rule,
-    input: struct {
-        owner_role: OwnerRole,
-        pool_manager_rule: AccessRule,
-        pool_contributor_rule: AccessRule,
-        resource_addresses: IndexSet<ResourceAddress>,
-        address_reservation: Option<GlobalAddressReservation>
-    },
-    output: type Global<MultiResourcePoolObjectTypeInfo>,
-    manifest_input: struct {
-        owner_role: OwnerRole,
-        pool_manager_rule: AccessRule,
-        pool_contributor_rule: AccessRule,
-        resource_addresses: IndexSet<ResourceAddress>,
-        address_reservation: Option<ManifestAddressReservation>
-    }
-}
-
-define_invocation! {
-    blueprint_name: MultiResourcePool,
     function_name: contribute,
     input: struct {
         buckets: Vec<Bucket>
