+++ radix-engine/src/blueprints/pool/v1/constants.rs
@@ -5,4 +5,3 @@ pub const MULTI_RESOURCE_POOL_BLUEPRINT_IDENT: &str = "MultiResourcePool";
 
 // Roles
 pub const POOL_MANAGER_ROLE: &str = "pool_manager_role";
-pub const POOL_CONTRIBUTOR_ROLE: &str = "pool_contributor_role";
