+++ radix-engine/src/utils/flash_state_generator.rs
@@ -1,9 +1,7 @@
 use crate::blueprints::models::KeyValueEntryContentSource;
 use crate::blueprints::package::*;
-use crate::blueprints::pool::v1::constants::*;
 use crate::blueprints::pool::v1::package::*;
 use crate::internal_prelude::*;
-use crate::system::attached_modules::role_assignment::*;
 use crate::system::system_db_reader::{ObjectCollectionKey, SystemDatabaseReader};
 use crate::track::{NodeStateUpdates, PartitionStateUpdates, StateUpdates};
 use crate::vm::VmApi;
@@ -17,7 +15,6 @@ use radix_engine_interface::api::ObjectModuleId;
 use radix_engine_interface::blueprints::consensus_manager::*;
 use radix_engine_interface::prelude::*;
 use radix_engine_interface::types::CollectionDescriptor;
-use radix_engine_store_interface::db_key_mapper::*;
 use radix_engine_store_interface::interface::*;
 use sbor::HasLatestVersion;
 use sbor::{generate_full_schema, TypeAggregator};
@@ -236,7 +233,6 @@ pub mod pools_package_v1_1 {
     ) -> StateUpdates {
         let mut state_updates = StateUpdates::default();
         generate_package_state_updated(db, &mut state_updates);
-        generate_role_assignment_update(db, &mut state_updates);
         state_updates
     }
 
@@ -313,73 +309,6 @@ pub mod pools_package_v1_1 {
                             .map(|(key, value)| (key, DatabaseUpdate::Set(value.into()))),
                     );
             })
-    }
-
-    /// Generates the state updates required for the new pool role.
-    ///
-    /// We have added a new role to the pool blueprints with v1.1 of the blueprints called the
-    /// `pool_contributor` role, which is the role that is allowed to contribute assets to the pool.
-    /// We would like pools that have already been instantiated to continue to function in the same
-    /// way as it did before the protocol update. As in, the protocol manager was the only role that
-    /// has contribution powers. So, for each pool that we find, we add a new role assignment for
-    /// the contributor role with the same rule as the pool manager. Thus, pools that were created
-    /// without this role will continue to function in the same way.
-    fn generate_role_assignment_update<S: SubstateDatabase + ListableSubstateDatabase>(
-        db: &S,
-        state_updates: &mut StateUpdates,
-    ) {
-        let reader = SystemDatabaseReader::new(db);
-
-        // Find all pools so that we apply this state changes to them
-        let pools = reader.partitions_iter().filter_map(|(node_id, _)| {
-            if node_id.entity_type().is_some_and(|entity_type| {
-                matches!(
-                    entity_type,
-                    EntityType::GlobalOneResourcePool
-                        | EntityType::GlobalTwoResourcePool
-                        | EntityType::GlobalMultiResourcePool
-                )
-            }) {
-                Some(node_id)
-            } else {
-                None
-            }
-        });
-
-        let key = SubstateKey::Map(
-            scrypto_encode(&ModuleRoleKey::new(ModuleId::Main, POOL_CONTRIBUTOR_ROLE)).unwrap(),
-        );
-        let mut already_seen = indexset! {};
-        for node_id in pools {
-            if already_seen.contains(&node_id) {
-                continue;
-            } else {
-                already_seen.insert(node_id)
-            };
-
-            let partition_number = reader
-                .get_partition_of_collection(
-                    &node_id,
-                    ModuleId::RoleAssignment,
-                    RoleAssignmentCollection::AccessRuleKeyValue.collection_index(),
-                )
-                .unwrap();
-
-            let substate_value = db
-                .get_substate(
-                    &SpreadPrefixKeyMapper::to_db_partition_key(&node_id, partition_number),
-                    &SpreadPrefixKeyMapper::to_db_sort_key(&SubstateKey::Map(
-                        scrypto_encode(&ModuleRoleKey::new(ModuleId::Main, POOL_MANAGER_ROLE))
-                            .unwrap(),
-                    )),
-                )
-                .unwrap();
-
-            state_updates
-                .of_node(node_id)
-                .of_partition(partition_number)
-                .update_substates([(key.clone(), DatabaseUpdate::Set(substate_value))])
-        }
     }
 
     struct MockVmApi;
